simplify ProgramInvite
